Removed unnecessary separation
