+++ src/Hypzeh.Web/ClientApp/components/routers/AppRouter.jsx
@@ -18,8 +18,8 @@ const AppRouter = () => (
     <AppMain>
       <Suspense fallback={<Loading />}>
         <Switch>
-          <Route exact path="/" component={About} />
-          <Route exact path="/about" component={Home} />
+          <Route exact path="/" component={Home} />
+          <Route exact path="/about" component={About} />
           <Route exact path="/projects" component={Projects} />
           <Route exact path="/contact" component={Contact} />
           <Route exact path="/error" render={({ history }) => <Error goBack={history.goBack} />} />
